Update v0.2.0
Added feature to check steam status. If anything other than normal, its considered possibly down, and won't add the track to broken or auto-skip, on a failure.
Fixed tracking and work with skips through /fs /forceskip and /skiplevel
Added Support to set Auto-Next level if the next track is known to be bad, this prevents ever loading into A05
If multiple tracks in a row are bad, skip to the next track that is not marked as bad.

Known Issue:
Small edge case, unable to handle commands that skip to a random level, if these load a broken map its not currently possible to detect what map was suppsoed to load